Add files via upload
(1) For Gaussian filtering, you need to generate a global mask for the filter (making sure the total weight of the mask is 1). For bilateral filtering, you need to generate a mask for each pixel based on its local neighborhood. Again, the weights need to sum to 1 for each mask.
(2) The neighborhood size for the mask may be adjusted but starting from a 5*5 would be a good idea.
(3) Two output images should be submitted: one for the Gaussian filtering and the other for the bilateral filtering, by using the sample image provided on Canvas.
(4) Please also submit the source code with your output images.